init (#10)

+++ humanresource/TABLE/emplagase.sql
@@ -1,5 +1,5 @@
 CREATE TABLE humans (
-  id   BIGINT,
+  id   INT,
   name VARCHAR(50),
   team VARCHAR(50),
   task VARCHAR(50),
